+++ modules/apps/site-navigation/site-navigation-test/src/testIntegration/java/com/liferay/site/navigation/service/test/SiteNavigationMenuServiceTest.java
@@ -32,7 +32,7 @@ import com.liferay.site.navigation.constants.SiteNavigationActionKeys;
 import com.liferay.site.navigation.constants.SiteNavigationConstants;
 import com.liferay.site.navigation.model.SiteNavigationMenu;
 import com.liferay.site.navigation.service.SiteNavigationMenuService;
-import com.liferay.site.navigation.util.SiteNavigationMenuTestUtil;
+import com.liferay.site.navigation.test.util.SiteNavigationMenuTestUtil;
 import com.liferay.site.navigation.util.comparator.SiteNavigationMenuCreateDateComparator;
 import com.liferay.site.navigation.util.comparator.SiteNavigationMenuNameComparator;
