+++ modules/apps/site-navigation/site-navigation-test/src/testIntegration/java/com/liferay/site/navigation/test/util/SiteNavigationMenuItemTestUtil.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.site.navigation.util;
+package com.liferay.site.navigation.test.util;
 
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.exception.PortalException;
